requirements.txt
streamlit
opencv-python
tensorflow
ultralytics
numpy
pillow
gdown
